WIP for parallel grouping sets.
For single rollup grouping sets, it is performed by aggregating in two
stages.  First, each process participating in the parallel portion of
the query performs a grouping sets aggregation step, producing a partial
result for each group of which that process is aware in each grouping
set. This is reflected in the plan as a Partial Aggregate node. Second,
the partial results are transferred to the leader via the Gather node.
Finally, the leader re-aggregates the results across all workers in
order to produce the final result. This is reflected in the plan as a
Finalize Aggregate node.

To prevent the Finalize Aggregate from combining partial results from
two different grouping sets incorrectly, we add a hidden gset_id column
to distinguish which tuple in the partial results comes from which
grouping set.

For grouping sets with multiple rollups, we generate a separate path for
each rollup and then append them for the final path.

Co-authored-by: Richard Guo <riguo@pivotal.io>
Co-authored-by: Paul Guo <pguo@pivotal.io>