Perl sript for temporary reference - delete when no longer needed
This perl script demonstrates how to start psql sessions in background
for running SQL commands that are expected to block.  It will serve as
a reference to write a TAP test that uses fault injectors to control
behavior of a remote PostgreSQL server (most likely a standby).  Some
of these faults are expected to cause blocking behavior when a SQL
command is run.

Caveat: it uses non-core module IPC::Run.  This may not be acceptable
but we shall see.