Add syntax to declare a step that is expected to block
The syntax is (note the "&" suffix):

    permutation "step_1" "step_2"& "step_3"

This is useful for defining steps that are expected to block for reasons
other than waiting on a lock.  For example, an injected fault may cause
a backend to suspend until another event occurs.  Isolation tester will
move on to execute step_3 without waiting for step_2 to finish, because
it is expected to block.

I've incorporated review feedback from Alvaro Herrera and
Michael Paquier.